test(specs): use mock server URL instead of github
